+++ lib/l10n/intl_en.arb
@@ -540,9 +540,10 @@
   "signTransactionListTransactionsHeader": "Transaction %1",
   "messengerTalkNotFound": "Messaging group not found",
   "messengerHeader": "Discussions",
-  "addMessengerGroup": "Create a messaging group",
+  "newTalk": "New discussion",
+  "newGroup": "New group",
   "addRemoteMessengerGroup": "Add this messaging group",
-  "addMessengerTalk": "Create a talk group",
+  "createGroup": "Create a group",
   "addMessengerGroupNameLabel": "Group name ...",
   "introNewMessengerTalk": "Choose a group name, and select people to invite",
   "messengerTalkAccessDescription": "Add or remove public keys that can access this talk group.",
