+++ lib/l10n/intl_fr.arb
@@ -517,9 +517,10 @@
 	"signTransactionListTransactionsHeader": "Transaction %1",
 	"messengerTalkNotFound": "Discussion introuvable",
 	"messengerHeader": "Discussions",
-	"addMessengerGroup": "Créer une discussion",
+	"newTalk": "Nouvelle discussion",
+	"newGroup": "Nouveau groupe",
 	"addRemoteMessengerGroup": "Ajouter cette discussion",
-	"addMessengerTalk": "Créer une conversation",
+	"createGroup": "Créer un groupe",
 	"addMessengerGroupNameLabel": "Nom du groupe ...",
 	"introNewMessengerTalk": "Choisissez le nom à donner à ce groupe, ainsi que les personnes à y inviter",
 	"messengerTalkAccessDescription": "Ajoutez ou supprimez des clés publiques pouvant accéder à cette conversation.",
